refactor(server): split fixed-slot reservation modules
