+++ apps/server/src/domain/reservations/index.ts
@@ -5,7 +5,9 @@ export * from "./repository/reservation-command.repository";
 export * from "./repository/reservation-query.repository";
 export * from "./services/cancel-reservation.service";
 export * from "./services/confirm-reservation.service";
-export * from "./services/fixed-slot.service";
+export * from "./services/fixed-slot/fixed-slot.helpers";
+export * from "./services/fixed-slot/fixed-slot.service";
+export type * from "./services/fixed-slot/fixed-slot.types";
 export * from "./services/reservation-command.service";
 export * from "./services/reservation-query.service";
 export * from "./services/reservation-stats.service";
